Добавлено: - открытие файлов на клиенте по двойному клику - отображение иконок файлов и папок - кнопка Создать папку - возможность работы с каталогами (пока только переименование, удаление) на клиенте и сервере - оключение кнопки Зарегестрироваться в окне регистрации после регистарции
Исправлено:
- тип с int на long переменных в FileMessage
- пересчет размера файлов (ранее округлялся до ближайщего целого)

+++ common/src/main/java/commands/FileMessageCommand.java
@@ -9,13 +9,13 @@ public class FileMessageCommand implements Serializable {
     private final String fileName;
     private final String destPath;
     private final long fileSize;
-    private final int partsOfFile;
+    private final long partsOfFile;
 
-    private int partNumber;
+    private long partNumber;
     private byte[] data;
 
 
-    public FileMessageCommand(String fileName, String destPath, long fileSize, int partsOfFile, int partNumber, byte[] data) {
+    public FileMessageCommand(String fileName, String destPath, long fileSize, long partsOfFile, long partNumber, byte[] data) {
         this.fileName = fileName;
         this.destPath = destPath;
         this.fileSize = fileSize;
@@ -44,15 +44,15 @@ public class FileMessageCommand implements Serializable {
         this.data = data;
     }
 
-    public int getPartsOfFile() {
+    public long getPartsOfFile() {
         return partsOfFile;
     }
 
-    public int getPartNumber() {
+    public long getPartNumber() {
         return partNumber;
     }
 
-    public void setPartNumber(int partNumber) {
+    public void setPartNumber(long partNumber) {
         this.partNumber = partNumber;
     }
 }
